Training Crisis or Not Model
train.py uses train.csv data to train a BERT model to determine if a tweet text is a about a crisis or not. The training took ~45 minutes on my ryzen 5800x. test.py uses sample text from test.csv as input to the trained model and outputs the model's determination.